add method to OrganizationRepository
